Mock missing comment pages in e2e test
